==> scripts/11_MLmetrics_figures.Rmd <==
---
title: "Figures ML metrics"
author: "Hannah Metzler"
date: "10/6/2021"
output: 
  pdf_document:
    df_print: kable
    keep_tex: true
url_colour: blue
always_allow_html: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE)
Sys.setlocale("LC_ALL", 'en_US.UTF-8')
options(scipen=99)
library(dplyr)
library(tidyr)
library(ggplot2)
library(ggpubr) #for multiple plots together

rm(list=ls())

# figure settings
cols = c("#332288", "#88CCEE", "#44AA99",  '#DDCC77', "#117733", '#999933') #
mycols2 = c("Awareness"= cols[1], "Suicide cases"=cols[2], "Prevention"=cols[3], "Suicidal"=cols[4], "Coping"=cols[5], "Irrelevant"=cols[6])
#text size
s=7
# s = 5 #for paper png figures JMIR, 15 within Rstudio or for pdfs
# change F1 to F subscript 1 in figures with facet_wrapt
facet_labels <- as_labeller(c(Precision = "Precision", Recall = "Recall",F1="F[1]"), default = label_parsed)
```

```{r, read & format data}

load('../results/intraclass_performance_6_classes.Rdata')

#order categories
ord_mcategory= c( "Suicidal", "Coping", "Awareness", "Prevention", "Suicide cases", "Irrelevant")

#function for formatting model intraclass performance scores for 6 categories (enter perClass$bert, perClass$xlnet, perClass$tfidf)
format6cl = function(df){
d6 =  df  %>% 
  mutate(Main_category = recode(main_category,"suicidality"="Suicidal", "coping"="Coping", "awareness"="Awareness", "prevention"="Prevention", "werther"="Suicide cases", "irrelevant"="Irrelevant")) %>% 
  #order factor levels
  mutate(Main_category = factor(Main_category, levels = ord_mcategory, labels = ord_mcategory)) %>% 
  select(-main_category) %>% 
  rename("Pr" = "Precision", "Re" = "Recall")

#long format with one column for score, and one each for Lower/Upper

#instructions here: 
# This regular expressions "(.*)(..)$" means:
# ( ) ( ) Look for two parts,
# (..) the first part should have just 2 characters
# (.*) the second part should have zero or more characters  at the “$” end of the string

longer<-pivot_longer(d6, cols=-"Main_category", names_pattern = "(..)(.*)$", names_to = c("Metric", "CI")) %>% 
  #if ci empty (because no second part), replace it with "value"
     mutate(CI=ifelse(CI=="", "Score" , CI))
#Create one column for each CI limit and score, keep the columns Metric and main_category
d6l <-pivot_wider(longer, id_cols = c(Metric, Main_category), names_from = c(CI), 
                     values_from = value,  names_repair = "check_unique") %>% 
    mutate(Metric = factor(Metric, levels = c("Pr", "Re", "F1"), labels = c("Precision", "Recall", "F1")))
return(d6l)
}

d6l = format6cl(perClass$bert)
```

# Figure 6 classes intra-class performance

```{r, BERT 2 intraclass 6 classes per class, cache=F}
plot6 = ggplot(d6l, aes(x=Main_category, y = Score, fill=Metric, colour = Metric)) + #dataset and variables to plot
  geom_bar(stat="identity", position = position_dodge2(width = 0.9))+
   labs(y="BERT performance score", x="") +
  scale_fill_manual(values=alpha(cols, 0.7), name="")+
  scale_colour_manual(values=cols, name="")+
  scale_y_continuous(breaks = seq(0, 1, 0.1), limits = c(0, 0.95))+
  theme_bw() +
  theme(legend.position="bottom", text=element_text(size=s), axis.title.x=element_blank(), plot.margin = margin(rep(0.5,4), unit="cm"))
plot6
```

```{r, BERT 2 intraclass 6 classes per metric, cache=T, fig.height=3, fig.width=6}


plotm = ggplot(d6l, aes(x=Main_category, y = Score, fill=Main_category, colour = Main_category)) + #dataset and variables to plot
  geom_bar(stat="identity", position = position_dodge2(width = 0.9))+
  labs(y="BERT performance", x="") +
  facet_wrap(~Metric, ncol=3, labeller = facet_labels) +
  scale_fill_manual(values=alpha(mycols2, 0.7), name="")+
   scale_colour_manual(values=mycols2, name="")+
  scale_y_continuous(breaks = seq(0, 1, 0.2), limits = c(0, 0.97))+
  theme_bw() +  
  theme(axis.text.x = element_text(angle = 60, hjust=1.2, vjust=1.15), legend.position="none", text=element_text(size=s), axis.title.x=element_blank()) #, plot.margin = margin(rep(0.4,4), unit="cm")
plotm
# ggsave('../figures/Figure_BERT_intraclass_6categories.pdf', plotm, width=6, height=3, dpi=300)
```


```{r, read data about suicide BERT}
load('../results/intraclass_performance_aboutsuicide.Rdata')
ds = perClass$bert
#long format
dsl = pivot_longer(ds, cols = c("Precision", "Recall", "F1"), names_to = "Metric", values_to ="Score") %>% 
  mutate(About_suicide = recode(about_suicide, correct = "About suicide", false="Off-topic")) %>% 
  select(-about_suicide) %>% 
  
  mutate(Metric = factor(Metric, levels = c("Precision", "Recall", "F1"), labels = c("Precision", "Recall", "F1"))) 
cols2 = c("#AA4371", "grey")
```
# Figure about suicide intra-class performance

```{r, BERT 2 intraclass about suicide, cache=T,  fig.height=3.5, fig.width=7}
plotabout = ggplot(dsl, aes(x=About_suicide, y = Score, fill=About_suicide, colour = About_suicide)) + #dataset and variables to plot
  geom_bar(stat="identity", position = position_dodge2(width = 0.9))+
  # geom_errorbar(aes(ymin = ci_low,ymax = ci_high), width=0.5, size=0.8) +
  labs(y="BERT performance", x="") +
  facet_wrap(~Metric, ncol=3, labeller = facet_labels) +
  scale_fill_manual(values=alpha(cols2, 0.7), name="")+
   scale_colour_manual(values=cols2, name="")+
  scale_y_continuous(breaks = seq(0, 1, 0.2), limits = c(0, 0.97))+
  theme_bw() +  
  theme(legend.position="none", text=element_text(size=s), axis.title.x=element_blank())#, plot.margin = margin(rep(0.4,4), unit="cm"))
plotabout
# ggsave('../figures/Figure_BERT_intraclass_about_suicide.pdf', plotabout, width=7, height=3.5, dpi=300)
```

# Combined figure for paper

```{r, fig.width=7, fig.height=6}

combined = ggarrange(plotm, plotabout, ncol=1, nrow=2, heights = c(1.3,1),
            labels = c("A", "B"), font.label = list(size = 8))
combined

ggsave('../figures/Figure1_BERT_intraclasses_permetric.png', combined, width=4, height=3.5, dpi = 300, units = "in", device='png')
#width*dpi = pixels, 4*300 and 3.5*300 is below or equal to 1200
#needed for png at JMIR: max  1200x1200 pixels, so reduce font size s to 7.5, and width/height to 4 and 3.5
```


# Model comparison 6 categories

Figure for presentations

```{r, read macro 6 categories data}
#not on new laptop yet, check hard drive
dma = read.csv('../results/Model_comparison_macro.csv', sep=',')
#long format
dmal = pivot_longer(dma, cols = c("Precision", "Recall", "F1", "Accuracy"), names_to = "Metric", values_to ="Score") %>% 
  mutate(Metric = factor(Metric, levels = c("Precision", "Recall", "F1", "Accuracy"), labels = c("Precision", "Recall", "F1", "Accuracy")), 
         Model = factor(Model, levels =c("Majority", "TF-IDF","BERT", "XLNet"), labels = c("Majority", "TF-IDF","BERT", "XLNet")))
```


```{r}
plotma = ggplot(dmal, aes(x=Model, y = Score, fill=Metric, colour = Metric)) + #dataset and variables to plot
  geom_bar(stat="identity", position = position_dodge2(width = 0.9))+
  # geom_errorbar(aes(ymin = ci_low,ymax = ci_high), width=0.5, size=0.8) +
  # labs(y="BERT performance", x="") +
  # facet_wrap(~Metric, ncol=2)+
  scale_fill_manual(values=alpha(cols, 0.7), name="")+
   scale_colour_manual(values=cols, name="")+
  scale_y_continuous(breaks = seq(0, 1, 0.2))+
  theme_bw() +  
  theme(legend.position="top", text=element_text(size=s), axis.title.x=element_blank(), plot.margin = margin(rep(0.4,4), unit="cm"))
plotma
ggsave('../figures/Model_comparison_6categories.pdf', plotma, width=6, height=4, dpi=300)
```


# Intraclass BERT 6 classes figure for presentations


```{r, BERT 2 intraclass 6 classes per metric, cache=T, fig.height=3, fig.width=8}
plotm = ggplot(d6l, aes(x=Main_category, y = Score, fill=Main_category, colour = Main_category)) + #dataset and variables to plot
  geom_bar(stat="identity", position = position_dodge2(width = 0.9))+
  labs(y="BERT performance", x="Category") +
  facet_wrap(~Metric, ncol=3)+
  scale_fill_manual(values=alpha(mycols2, 0.7), name="",  guide = guide_legend(nrow=1, ncol = 6, byrow = T))+
   scale_colour_manual(values=mycols2, name="", guide = guide_legend(nrow=1, ncol = 6, byrow = T))+
  scale_y_continuous(breaks = seq(0, 1, 0.2), limits = c(0, 0.97))+
  theme_bw() +  
  theme(axis.text.x = element_blank(), legend.position="bottom", text=element_text(size=s), plot.margin = margin(rep(0.4,4), unit="cm"),   axis.ticks.x = element_blank())
plotm
ggsave('../figures/Figure_BERT_intraclass_6categories_legendbottom.pdf', plotm, width=8, height=3, dpi=300)
```
